requirements.txt
pip==20.2.4
absl-py==0.11.0
astor==0.8.1
autopep8==1.5.4
cached-property==1.5.2
click==7.1.2
flake8==3.8.4
Flask==1.1.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==3.1.0
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
numpy==1.19.4
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pycodestyle==2.6.0
pyflakes==2.2.0
six==1.15.0
tensorboard==1.15.0
https://files.pythonhosted.org/packages/02/36/9a02e27f0ec248b676a380ffe910c1858e3af3027c0d4d513dd0b56a5613/tensorflow-1.15.3-cp36-cp36m-manylinux2010_x86_64.whl
tensorflow-estimator==1.15.1
termcolor==1.1.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0

